requirements.txt
pandas
numpy
matplotlib
streamlit
requests
beautifulsoup4
transformers
torch
scikit-learn
ta
openai
python-dotenv
